+++ rasa/core/trackers.py
@@ -112,6 +112,7 @@ class DialogueStateTracker(object):
         self.latest_action_name = None
         # Stores the most recent message sent by the user
         self.latest_message = None
+        self.action_utterances = []
         self.latest_bot_utterance = None
         self._reset()
         self.active_form = {}
@@ -189,8 +190,7 @@ class DialogueStateTracker(object):
             # reset form rejection if it was predicted again
             self.active_form["rejected"] = False
 
-    def current_slot_values(self):
-        # type: () -> Dict[Text, Any]
+    def current_slot_values(self) -> [Dict[Text, Any]]:
         """Return the currently set values of the slots"""
         return {key: slot.value for key, slot in self.slots.items()}
 
@@ -227,8 +227,7 @@ class DialogueStateTracker(object):
         """State whether the tracker is currently paused."""
         return self._paused
 
-    def idx_after_latest_restart(self):
-        # type: () -> int
+    def idx_after_latest_restart(self) -> int:
         """Return the idx of the most recent restart in the list of events.
 
         If the conversation has not been restarted, ``0`` is returned."""
@@ -239,8 +238,7 @@ class DialogueStateTracker(object):
                 idx = i + 1
         return idx
 
-    def events_after_latest_restart(self):
-        # type: () -> List[Event]
+    def events_after_latest_restart(self) -> List[Event]:
         """Return a list of events after the most recent restart."""
         return list(self.events)[self.idx_after_latest_restart() :]
 
@@ -252,6 +250,13 @@ class DialogueStateTracker(object):
         return DialogueStateTracker(
             UserMessage.DEFAULT_SENDER_ID, self.slots.values(), self._max_event_history
         )
+
+    def log_action_utterances_to_events(self) -> None:
+        """Logs the utterances collected during an action to events."""
+        if self.action_utterances:
+            for utter_event in self.action_utterances:
+                self.update(utter_event)
+            self.action_utterances = []
 
     def generate_all_prior_trackers(self):
         # type: () -> Generator[DialogueStateTracker, None, None]
@@ -352,8 +357,7 @@ class DialogueStateTracker(object):
                 applied_events.append(event)
         return applied_events
 
-    def replay_events(self):
-        # type: () -> None
+    def replay_events(self) -> None:
         """Update the tracker based on a list of events."""
 
         applied_events = self.applied_events()
@@ -398,8 +402,7 @@ class DialogueStateTracker(object):
 
         return tracker  # yields the final state
 
-    def as_dialogue(self):
-        # type: () -> Dialogue
+    def as_dialogue(self) -> Dialogue:
         """Return a ``Dialogue`` object containing all of the turns.
 
         This can be serialised and later used to recover the state
@@ -484,8 +487,7 @@ class DialogueStateTracker(object):
     # only be called by events, not directly. Rather update the tracker
     # with an event that in its ``apply_to`` method modifies the tracker.
     ###
-    def _reset(self):
-        # type: () -> None
+    def _reset(self) -> None:
         """Reset tracker to initial state - doesn't delete events though!."""
 
         self._reset_slots()
@@ -496,8 +498,7 @@ class DialogueStateTracker(object):
         self.followup_action = ACTION_LISTEN_NAME
         self.active_form = {}
 
-    def _reset_slots(self):
-        # type: () -> None
+    def _reset_slots(self) -> None:
         """Set all the slots to their initial value."""
 
         for slot in self.slots.values():
@@ -535,8 +536,7 @@ class DialogueStateTracker(object):
 
         self.followup_action = action
 
-    def clear_followup_action(self):
-        # type: () -> None
+    def clear_followup_action(self) -> None:
         """Clears follow up action when it was executed."""
 
         self.followup_action = None