+++ tests/core/test_channels.py
@@ -27,6 +27,40 @@ def fake_sanic_run(*args, **kwargs):
     logger.info("Rabatnic: Take this and find Sanic! I want him here by supper time.")
 
 
+async def test_send_response(default_channel, default_tracker):
+    text_only_message = {"text": "hey"}
+    image_only_message = {"image": "https://i.imgur.com/nGF1K8f.jpg"}
+    text_and_image_message = {
+        "text": "look at this",
+        "image": "https://i.imgur.com/T5xVo.jpg",
+    }
+
+    await default_channel.send_response(default_tracker.sender_id, text_only_message)
+    await default_channel.send_response(default_tracker.sender_id, image_only_message)
+    await default_channel.send_response(
+        default_tracker.sender_id, text_and_image_message
+    )
+    collected = default_channel.messages
+
+    assert len(collected) == 4
+
+    # text only message
+    assert collected[0] == {"recipient_id": "my-sender", "text": "hey"}
+
+    # image only message
+    assert collected[1] == {
+        "recipient_id": "my-sender",
+        "image": "https://i.imgur.com/nGF1K8f.jpg",
+    }
+
+    # text & image combined - will result in two messages
+    assert collected[2] == {"recipient_id": "my-sender", "text": "look at this"}
+    assert collected[3] == {
+        "recipient_id": "my-sender",
+        "image": "https://i.imgur.com/T5xVo.jpg",
+    }
+
+
 async def test_console_input():
     from rasa.core.channels import console